Fix type mismatch : Signed to Unsigned.
Integer variable i is used for counting iteration number. Since return
type of functions is unsigned long, the variable type is casted to
unsigned long at the return stage.

+++ Common/src/matrix_structure.cpp
@@ -1189,7 +1189,7 @@ unsigned long CSysMatrix::Jacobi_Smoother(const CSysVector & b, CSysVector & x, 
     cout << "# Iteration = " << i << ": |res|/|res0| = "  << norm_r/norm0 << ".\n" << endl;
   }
   
-  return i;
+  return (unsigned long) i;
   
 }
 
@@ -1493,7 +1493,7 @@ unsigned long CSysMatrix::ILU0_Smoother(const CSysVector & b, CSysVector & x, CM
     cout << "# Iteration = " << i << ": |res|/|res0| = "  << norm_r/norm0 << ".\n" << endl;
   }
   
-  return i;
+  return (unsigned int) i;
   
 }
 
@@ -1656,7 +1656,7 @@ unsigned long CSysMatrix::LU_SGS_Smoother(const CSysVector & b, CSysVector & x, 
     cout << "# Iteration = " << i << ": |res|/|res0| = "  << norm_r/norm0 << ".\n" << endl;
   }
   
-  return i;
+  return (unsigned int) i;
   
 }
